chore: Plumbs noir subrepo into yarn-project. (#3420)
For a less noisey review:
https://github.com/AztecProtocol/aztec-packages/pull/3427
It contains everything prior to the noir pulls. This is PR we actually
want to merge though.


* Adds `noir-packages` job to build the ts packages.
* Fix `sed` use in bb.js to work on Mac.
* Bootstrap scripts for noir, call it from bootstrap.sh.
* Fix to `build_local`.
* Use the noir bootstrap scripts to build the docker containers.
* `yarn-project` now builds the protocol circuits from the noir build in
bootstrap and dockerfile.
* `yarn-project` portals the noir packages in. **What does this mean for
releases!?*
* Fixes as we're now on noir 19.
* Remove `noir-protocol-circuits/src/target` from repo as they're now
built.
* Bootstrap improvements:
  * Defaults to a "lighter" bootstrap.
  * `bootstrap.sh clean` resets repo to as if fresh clone. (be careful).
* Introduces ability to test the bootstrap process via a docker
container. (might add a nightly job, plus alternative ubuntu releases).

+++ barretenberg/cpp/scripts/install-wasi-sdk.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -eu
 
+if [[ -d ./src/wasi-sdk-20.0 && -d ./src/wasi-sdk-20.0+threads ]]; then
+  echo "WASI already installed. Skipping."
+  exit 0
+fi
+
 # Clean.
 rm -rf ./src/wasi-sdk-*
 
@@ -17,15 +22,15 @@ fi
 # Install the webassembly toolchain.
 mkdir -p src
 cd ./src
-curl -s -L https://github.com/WebAssembly/wasi-sdk/releases/download/wasi-sdk-20/wasi-sdk-20.0-$OS.tar.gz | tar zxfv -
+curl -s -L https://github.com/WebAssembly/wasi-sdk/releases/download/wasi-sdk-20/wasi-sdk-20.0-$OS.tar.gz | tar zxf -
 if [ "$OS" == "linux" ]; then
   # Exceptional linux case that provides an llvm build that works on Ubuntu 20.
-  curl -s -L https://wasi-sdk.s3.eu-west-2.amazonaws.com/yamt-wasi-sdk-20.0.threads.tgz | tar zxfv -
+  curl -s -L https://wasi-sdk.s3.eu-west-2.amazonaws.com/yamt-wasi-sdk-20.0.threads.tgz | tar zxf -
 else
   # For other operating systems, first download the standard release (this is to get the llvm build).
-  curl -s -L https://github.com/WebAssembly/wasi-sdk/releases/download/wasi-sdk-20%2Bthreads/wasi-sdk-20.0.threads-$OS.tar.gz | tar zxfv -
+  curl -s -L https://github.com/WebAssembly/wasi-sdk/releases/download/wasi-sdk-20%2Bthreads/wasi-sdk-20.0.threads-$OS.tar.gz | tar zxf -
   # Replace wasi-sysroot in wasi-sdk-20.0+threads with our custom build.
   # It contains libc++ and a patch by yamt to improve thread join stability.
   # Can remove once future releases are more stable.
-  curl -s -L https://wasi-sdk.s3.eu-west-2.amazonaws.com/yamt-wasi-sysroot-20.0.threads.tgz | tar zxfv -
+  curl -s -L https://wasi-sdk.s3.eu-west-2.amazonaws.com/yamt-wasi-sysroot-20.0.threads.tgz | tar zxf -
 fi
